feat(backend): restructure projects into Api, Application, Domain, Infrastructure, Tests layers
