+++ backend/RubberDev.Application/Interfaces/ICartoonCharacterService.cs
@@ -1,6 +1,6 @@
-﻿using RubberDev.Models;
+﻿using RubberDev.Domain.Entities;
 
-namespace RubberDev.Services.Interfaces;
+namespace RubberDev.Application.Interfaces;
 
 public interface ICartoonCharacterService
 {
