+++ backend/RubberDev.Application/Services/CartoonCharacterService.cs
@@ -1,32 +1,31 @@
-﻿using RubberDev.Models;
-using RubberDev.Brokers;
-using RubberDev.Services.Interfaces;
+﻿using RubberDev.Application.Interfaces;
+using RubberDev.Domain.Entities;
 
-namespace RubberDev.Services;
+namespace RubberDev.Application.Services;
 
 public class CartoonCharacterService : ICartoonCharacterService
 {
-    private readonly IStorageBroker storageBroker;
+    private readonly IStorageBroker _storageBroker;
 
     public CartoonCharacterService(IStorageBroker storageBroker) =>
-        this.storageBroker = storageBroker;
+        this._storageBroker = storageBroker;
 
     public async ValueTask<CartoonCharacter> AddCartoonCharacterAsync(CartoonCharacter character)
     {
         // TODO: add any business validations here
-        return await this.storageBroker.InsertCartoonCharacterAsync(character);
+        return await this._storageBroker.InsertCartoonCharacterAsync(character);
     }
 
     public IQueryable<CartoonCharacter> RetrieveAllCartoonCharacters() =>
-        this.storageBroker.SelectAllCartoonCharacters();
+        this._storageBroker.SelectAllCartoonCharacters();
 
     public async ValueTask<CartoonCharacter> RetrieveCartoonCharacterByIdAsync(Guid characterId) =>
-        await this.storageBroker
+        await this._storageBroker
             .SelectCartoonCharacterByIdAsync(characterId);
 
     public async ValueTask<CartoonCharacter> ModifyCartoonCharacterAsync(CartoonCharacter character) =>
-        await this.storageBroker.UpdateCartoonCharacterAsync(character);
+        await this._storageBroker.UpdateCartoonCharacterAsync(character);
 
     public async ValueTask<CartoonCharacter> RemoveCartoonCharacterAsync(Guid characterId) =>
-        await this.storageBroker.DeleteCartoonCharacterAsync(characterId);
+        await this._storageBroker.DeleteCartoonCharacterAsync(characterId);
 }
