+++ backend/src/RubberDev.Tests/RubberDev.Tests.csproj
@@ -23,4 +23,9 @@
     </PackageReference>
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Application.Tests\" />
+    <Folder Include="Infrastructure.tests\" />
+  </ItemGroup>
+
 </Project>
